python_gcp_functions/genDefaultLevelsEndpoint/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.9.0
colorama==0.4.6
db-dtypes==1.1.1
deprecation==2.1.0
Flask==2.3.3
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
pandas-gbq==0.19.2
proto-plus==1.22.3
protobuf==4.24.3
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydata-google-auth==1.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
urllib3==1.26.16
watchdog==1.0.2
Werkzeug==2.3.7
yarl==1.9.2
